==> Twitter Gender Prediction/1-credentials.py <==
#CONSUMER_KEY = 'kvfzUYlmquxM4xtE2W3rHHKqe'
#CONSUMER_SECRET = 'nEjmmV3alvj2Hi7z1rwJdHtAX19ULBJGBI4APblXxgZAfyqkF3'
#ACCESS_KEY = '867050392521134084-J2PjZtFxJXpMbDyLnSnqzmGPIylcnvI'
#ACCESS_SECRET = 'iTq0fbGRTcpW4SGlMAXrhwO9r6nGw7Gcb6wp5A4NaxSgE'

ACCESS_KEY = "1050469047391973382-olF5OR65CdtigUA8YL7HCoEDiGv9kw"
ACCESS_SECRET = "D5nJgmrEiNaU2FYXcO5aBddW1J5xUwgk0YX8J8QhnSKCx"
CONSUMER_KEY = "pDxBEeRDSkHxLWlNXtqXtBft5"
CONSUMER_SECRET = "YmIUBGJKfp8Mf3nbQYKBnTxlJlsuU3xVRKxDUmqtbxeimg41y2"
